+++ src/server.h
@@ -55,7 +55,7 @@ struct server
 {
     int fd;
     int stage;
-    int buf_len;
+    ssize_t buf_len;
     int buf_idx;
     char *buf; // server send from, remote recv into
     asyncns_query_t *query;
@@ -77,7 +77,7 @@ struct remote_ctx
 struct remote
 {
     int fd;
-    int buf_len;
+    ssize_t buf_len;
     int buf_idx;
     char *buf; // remote send from, server recv into
     struct remote_ctx *recv_ctx;
