+++ src/tunnel.h
@@ -52,7 +52,7 @@ struct server_ctx
 struct server
 {
     int fd;
-    int buf_len;
+    ssize_t buf_len;
     int buf_idx;
     char *buf; // server send from, remote recv into
     struct enc_ctx *e_ctx;
@@ -74,7 +74,7 @@ struct remote_ctx
 struct remote
 {
     int fd;
-    int buf_len;
+    ssize_t buf_len;
     int buf_idx;
     char *buf; // remote send from, server recv into
     struct remote_ctx *recv_ctx;
